update fillContent to enable future development
rewrote the fillContent function to be more readable by not calling
the dom-util function. Included the index values from the global
projectList constant as a html dataset value. I'll use this later
to connect 'edit' and 'delete' btns. Same is true for the 'ToDo'
attributes.

Within the dom-utils, addToContainer remains useful but
createNewContainer may be too annoying to read when used.